fix: resolve Kotlin compile errors and add Gradle wrapper
- MainActivity.setupNav now takes savedInstanceState explicitly
- TimerService uses CATEGORY_PROGRESS (CATEGORY_TIMER does not exist)
- TimeFormat.dayDelta coerces the cross-year branch to Int
- Remove deprecated android.defaults.buildfeatures.buildconfig (already in build.gradle.kts)
- Add gradlew, gradlew.bat, gradle-wrapper.jar

+++ app/src/main/java/com/futureclock/app/util/TimeFormat.kt
@@ -78,7 +78,7 @@ object TimeFormat {
         val zoneDay = zoneCal.get(Calendar.DAY_OF_YEAR)
         val zoneYear = zoneCal.get(Calendar.YEAR)
         return if (zoneYear == deviceYear) zoneDay - deviceDay
-        else (zoneCal.timeInMillis - deviceCal.timeInMillis) / (24L * 60 * 60 * 1000)
+        else ((zoneCal.timeInMillis - deviceCal.timeInMillis) / (24L * 60 * 60 * 1000)).toInt()
     }
 }
